backend/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
python-multipart==0.0.6
numpy==1.23.4
torch==2.0.0
torchvision==0.15.0
facenet-pytorch==2.5.3
Pillow==9.5.0
python-jose==3.3.0
